% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasalClassicalHeatmap.R
\docType{package}
\encoding{UTF-8}
\name{BasalClassicalHeatmap-package}
\alias{BasalClassicalHeatmap-package}
\alias{BasalClassicalHeatmap}
\title{BasalClassicalHeatmap: TODO}
\description{
To learn more about \pkg{BasalClassicalHeatmap} package see:
\url{https://github.com/adeschen/BasalClassicalHeatmap/wiki}
}
\seealso{
\itemize{
\item \code{\link{createHeatmap}} {for creating a heating using a 
gene list and RNA-seq expression data.}
}
}
\author{
Maintainer:
Astrid Deschênes <adeschen@hotmail.com>
}
\keyword{package}
