% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasalClassicalHeatmap.R
\docType{data}
\name{demo_PDAC_TCGA}
\alias{demo_PDAC_TCGA}
\title{TODO}
\format{
a \code{data.frame} containing normalized expected counts for genes 
associated to Basal-like/Classical classification. The rows are containing 
normalized expected counts. The \code{data.frame} 
contains one column per sample in addition to a column called 'GENE' that 
contains the name of the genes represented in each row.
}
\usage{
data(demo_PDAC_TCGA)
}
\description{
TODO
}
\examples{

## Load Miyabayashi et al 2020 gene list used for 
## Basal-like/Classical classification
data(Miyabayashi_2020)

## Load the expression dataset
data(demo_PDAC_TCGA)

## Create a heatmap using the Miyabayashi et al 2020 gene list
createHeatmap(gene_list=Miyabayashi_2020, rna_data=demo_PDAC_TCGA, 
     gene_column="GENE")
     
}
\references{
The TCGA Research Network: https://www.cancer.gov/tcga
}
\seealso{
\itemize{
\item \code{\link{createHeatmap}} { for creating a heating using a 
gene list and RNA-seq expression data.}
}
}
\keyword{datasets}
