% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureHeatmap.R
\docType{data}
\name{Miyabayashi_2020}
\alias{Miyabayashi_2020}
\title{The Basal-like and Classical gene list used in Miyabayashi et al 2020.}
\format{
a \code{data.frame} with the following columns:
\itemize{
\item \code{SYMBOL} { a \code{character} string representing the
symbol of the gene. }
\item \code{Class} { a \code{character} string representing the class 
(Basal-like or Classical) associated to the gene. }
}
}
\usage{
data(Miyabayashi_2020)
}
\description{
The Basal-like and Classical gene list used in Miyabayashi et al 2020.
}
\examples{

## Load Miyabayashi et al 2020 gene list used for 
## Basal-like/Classical classification
data(Miyabayashi_2020)

## The dataset contains 2 columns
## The first column has the gene symbols
## The second column the class assignation (Basal-like or Classical)
head(Miyabayashi_2020)

}
\references{
Miyabayashi K, Baker LA, Deschênes A, Traub B, Caligiuri G, Plenker D, 
Alagesan B, Belleau P, Li S, Kendall J, Jang GH, Kawaguchi RK, 
Somerville TDD, Tiriac H, Hwang CI, Burkhart RA, Roberts NJ, Wood LD, 
Hruban RH, Gillis J, Krasnitz A, Vakoc CR, Wigler M, Notta F, 
Gallinger S, Park Y, Tuveson DA. Intraductal Transplantation Models of 
Human Pancreatic Ductal Adenocarcinoma Reveal Progressive Transition of 
Molecular Subtypes. Cancer Discov. 2020 Oct;10(10):1566-1589. 
doi: 10.1158/2159-8290.CD-20-0133. 
Epub 2020 Jul 23. PMID: 32703770; PMCID: PMC7664990.
}
\seealso{
\itemize{
\item \code{\link{createHeatmap}} { for creating a heating using a 
gene list and RNA-seq expression data.}
}
}
\keyword{datasets}
