% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_internal.R
\encoding{UTF-8}
\name{prepareExpression}
\alias{prepareExpression}
\title{Prepare the expression data by selecting basal-like/classical genes
and running normalization}
\usage{
prepareExpression(rna_data, gene_list, gene_column = "GENE")
}
\arguments{
\item{rna_data}{a \code{data.frame} containing the expression values for 
the studied samples. There must be one column per sample and 
an extra column for the gene names. The name of the column containing the 
gene names is passed in the \code{gene_column} parameter.}

\item{gene_list}{a \code{data.frame} containing the gene list used for the
basal-like/classical classification.}

\item{gene_column}{a \code{character} string representing the name of the
column in the \code{rna_data} parameter that contains the gene names.}
}
\value{
a \code{list} containing 2 entries: \itemize{
\item{DATA} { a \code{data.frame} containing the normalized expression 
values for the genes that are part of the classical/basal-like gene list. 
The gene names are used as row names and the gene order fits the order 
of the gene names in the \code{ROW_INFO} \code{data.frame}.}
\item{ROW_INFO} {a \code{data.frame} containing information about the genes 
from basal-like/classical classification that are present in the 
\code{DATA} \code{data.frame}. The gene names are used as row names and the 
gene order fits the order of the gene names in 
the \code{DATA} \code{data.frame}.}
}
}
\description{
The function selects the expression values related to the 
gene list passed in argument. Then, the function adds one to the expression 
values before log transforming the values. Then, the function normalizes 
the data so that each row has mean 0 and variance 1.
}
\examples{

## Load Moffitt et al 2015 gene list
data("Moffitt_2015")

## Load demo PDAC TCGA expression data
data("demo_PDAC_TCGA")

## Prepare the data for the heatmap creation step
result <- BasalClassicalHeatmap:::prepareExpression(rna_data=demo_PDAC_TCGA, 
    gene_list=Moffitt_2015, gene_column="GENE")

}
\author{
Astrid Deschênes
}
\keyword{internal}
