% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureHeatmap.R
\docType{package}
\encoding{UTF-8}
\name{SignatureHeatmap-package}
\alias{SignatureHeatmap-package}
\alias{SignatureHeatmap}
\title{SignatureHeatmap: TODO}
\description{
To learn more about \pkg{SignatureHeatmap} package see:
\url{https://github.com/adeschen/SignatureHeatmap/wiki}
}
\seealso{
\itemize{
\item \code{\link{createHeatmap}} {for creating a heating using a 
gene signature list and RNA-seq expression data.}
}
}
\author{
Maintainer:
Astrid Deschênes <adeschen@hotmail.com>
}
\keyword{package}
