% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasalClassicalHeatmap.R
\docType{data}
\name{Tiriac_2018}
\alias{Tiriac_2018}
\title{The Basal-like and Classical gene list used in Tiriac et al 2018.}
\format{
a \code{data.frame} with the following columns:
\itemize{
\item \code{SYMBOL} { a \code{character} string representing the
symbol of the gene. }
\item \code{Class} { a \code{character} string representing the class 
(Basal-like or Classical) associated to the gene. }
}
}
\usage{
data(Tiriac_2018)
}
\description{
The Basal-like and Classical gene list used in Tiriac et al 2018.
}
\examples{

## Load Tiriac et al 2018 gene list used for 
## Basal-like/Classical classification
data(Tiriac_2018)

## The dataset contains 2 columns
## The first column has the gene symbols
## The second column the class assignation (Basal-like or Classical)
head(Tiriac_2018)


}
\seealso{
\itemize{
\item \code{\link{createHeatmap}} { for creating a heating using a 
gene list and RNA-seq expression data.}
}
}
\keyword{datasets}
