% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_internal.R
\encoding{UTF-8}
\name{validateCreateHeatmap}
\alias{validateCreateHeatmap}
\title{Validate the parameters for createHeatmap() function}
\usage{
validateCreateHeatmap(
  gene_list,
  rna_data,
  gene_column,
  name,
  show_column_dend,
  show_row_dend,
  cluster_columns,
  cluster_rows,
  log_data
)
}
\arguments{
\item{gene_list}{a \code{data.frame} containing the gene list used for the
basal-like/classical classification. The \code{data.frame} must contain 
two columns: \code{SYMBOL} and \code{Class}. The \code{SYMBOL} contains the 
gene symbol while the \code{Class} column contains its classification. The 
classification can be \code{Classical} or \code{Basal-like}.}

\item{rna_data}{\code{data.frame} containing the expression level for the 
samples. Each column must represent a sample. An extra column for the 
gene name must be present. The name of this column must correspond to the 
parameter \code{gene_column}.}

\item{gene_column}{a \code{character} string representing the name of the
column in the \code{rna_data} parameter that contains the gene names. 
Default: \code{GENE}.}

\item{name}{a \code{character} string representing the title of the 
heatmap legend.}

\item{show_column_dend}{a code{logical} indicating if the column 
dendrogram should be shown.}

\item{show_row_dend}{a code{logical} indicating if the row
dendrogram should be shown.}

\item{cluster_columns}{a code{logical} indicating if the columns
should be clustered.}

\item{cluster_rows}{a code{logical} indicating if the rows
should be clustered.}

\item{log_data}{a code{logical} indicating if the expression data
should be log transformed.}
}
\value{
\code{0L} when the function is successful.
}
\description{
This function validates the parameters input in the 
createHeatmap() function.
}
\examples{

## Load Moffitt et al 2015 gene list
data("Moffitt_2015")


## Load TCGA PDAC dataset
data("demo_PDAC_TCGA")

## TODO
BasalClassicalHeatmap:::validateCreateHeatmap(gene_list=Moffitt_2015,
    rna_data=demo_PDAC_TCGA, gene_column="GENE", name="Expression", 
    show_column_dend=TRUE, show_row_dend=TRUE, cluster_columns=TRUE,
    cluster_rows=TRUE, log_data=FALSE)


}
\author{
Astrid Deschênes
}
\keyword{internal}
