% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap.R
\encoding{UTF-8}
\name{createHeatmap}
\alias{createHeatmap}
\title{Generate a heatmap based on transcriptomic data for a 
basal-like/classical gene list}
\usage{
createHeatmap(
  gene_list,
  rna_data,
  gene_column = "GENE",
  name = "Expression",
  clustering_distance_rows = c("euclidean", "maximum", "manhattan", "canberra",
    "binary", "minkowski", "pearson", "spearman", "kendall"),
  clustering_distance_columns = c("euclidean", "maximum", "manhattan", "canberra",
    "binary", "minkowski", "pearson", "spearman", "kendall"),
  show_column_dend = TRUE,
  show_row_dend = TRUE,
  cluster_columns = TRUE,
  cluster_rows = TRUE,
  log_data = TRUE,
  heatmap_legend_param = gpar(title_gp = gpar(col = "black", fontsize = 11, fontface =
    "bold")),
  ...
)
}
\arguments{
\item{gene_list}{a \code{data.frame} containing the gene list used for the
basal-like/classical classification. The \code{data.frame} must contain 
two columns: \code{SYMBOL} and \code{Class}. The \code{SYMBOL} contains the 
gene symbol while the \code{Class} column contains its classification. The 
classification can be \code{Classical} or \code{Basal-like}.}

\item{rna_data}{\code{data.frame} containing the expression level for the 
samples. Each column must represent a sample. An extra column for the 
gene name must be present. The name of this column must correspond to the 
parameter \code{gene_column}.}

\item{gene_column}{a \code{character} string representing the name of the
column in the \code{rna_data} parameter that contains the gene names. 
Default: \code{GENE}.}

\item{name}{a \code{character} string representing the title of the 
heatmap legend. Default: \code{"Expression"}.}

\item{clustering_distance_rows}{a \code{character} string representing 
the method uses to cluster the rows of the heatmap. 
The available choices are: 
"euclidean", "maximum", "manhattan", "canberra", 
"binary", "minkowski", "pearson", "spearman" and "kendall". 
 Default: \code{"euclidean"}.}

\item{clustering_distance_columns}{a \code{character} string representing 
the method uses to cluster the columns of the heatmap.
The available choices are: 
"euclidean", "maximum", "manhattan", "canberra", 
"binary", "minkowski", "pearson", "spearman" and "kendall". 
Default: \code{"euclidean"}.}

\item{show_column_dend}{a code{logical} indicating if the column 
dendrogram should be shown. Default: \code{TRUE}.}

\item{show_row_dend}{a code{logical} indicating if the row
dendrogram should be shown. Default: \code{TRUE}.}

\item{cluster_columns}{a code{logical} indicating if the columns
should be clustered. Default: \code{TRUE}.}

\item{cluster_rows}{a code{logical} indicating if the rows
should be clustered. Default: \code{TRUE}.}

\item{log_data}{a \code{logical} indicating if the data should be log 
transformed. If so, the data will be log2(value + 1) transformed. 
Default: \code{TRUE}.}

\item{heatmap_legend_param}{an object of class '\code{gpar}'. Default: 
\code{gpar(title_gp=gpar(col="black", fontsize=11, fontface="bold"))}.}

\item{\ldots}{further arguments passed to ComplexHeatmap::Heatmap() function.}
}
\value{
an object of class {ComplexHeatmap} that represent a heatmap based 
on transcriptomic data for a specific basal-like/classical gene list.
}
\description{
The function selects the expression values related to the 
gene list passed in argument. Then, the function adds one to the expression 
values before log transforming the values. Then, the function normalizes 
the data so that each row has mean 0 and variance 1. Finally, the function 
generates a heatmap using input parameters.
}
\examples{

## Load Moffitt et al 2015 gene list
data("Moffitt_2015")

## Load demo PDAC TCGA expression data
data("demo_PDAC_TCGA")

## Create heatmap with default parameters
## The column in the rna_data data.frame is the default name "GENE"
createHeatmap(gene_list=Moffitt_2015, rna_data=demo_PDAC_TCGA)

}
\seealso{
The default method ComplexHeatmap::Heatmap().
}
\author{
Astrid Deschênes
}
