% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasalClassicalHeatmap.R
\docType{data}
\name{demo_PDAC_TCGA}
\alias{demo_PDAC_TCGA}
\title{The Basal-like and Classical gene list used in Miyabayashi et al 2020.}
\format{
a \code{data.frame} with the following columns:
\itemize{
\item \code{SYMBOL} { a \code{character} string representing the
symbol of the gene. }
\item \code{Class} { a \code{character} string representing the class 
(Basal-like or Classical) associated to the gene. }
}
}
\usage{
data(demo_PDAC_TCGA)
}
\description{
The Basal-like and Classical gene list used in Miyabayashi et al 2020.
}
\examples{

## Load Miyabayashi et al 2020 gene list used for 
## Basal-like/Classical classification
data(Miyabayashi_2020)

data(demo_PDAC_TCGA)

createHeatmap(gene_list=Miyabayashi_2020, rna_data=demo_PDAC_TCGA, 
     gene_column="GENE")
}
\seealso{
\itemize{
\item \code{\link{createHeatmap}} { for creating a heating using a 
gene list and RNA-seq expression data.}
}
}
\keyword{datasets}
