% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasalClassicalHeatmap.R
\docType{data}
\name{Moffitt_2015}
\alias{Moffitt_2015}
\title{The Basal-like and Classical gene list used in Moffitt et al 2015.}
\format{
a \code{data.frame} with the following columns:
\itemize{
\item \code{SYMBOL} { a \code{character} string representing the
symbol of the gene. }
\item \code{Class} { a \code{character} string representing the class 
(Basal-like or Classical) associated to the gene. }
}
}
\usage{
data(Moffitt_2015)
}
\description{
The Basal-like and Classical gene list used in Moffitt et al 2015.
}
\examples{

## Load Moffitt et al 2015 gene list used for 
## Basal-like/Classical classification
data(Moffitt_2015)

## The dataset contains 2 columns
## The first column has the gene symbols
## The second column the class assignation (Basal-like or Classical)
head(Moffitt_2015)

}
\seealso{
\itemize{
\item \code{\link{createHeatmap}} { for creating a heating using a 
gene list and RNA-seq expression data.}
}
}
\keyword{datasets}
